perf(web): shrink the critical startup graph
