+++ apps/web/src/routes/-ExistingProjectIndex.tsx
@@ -1,0 +1,60 @@
+import { scopeProjectRef } from "@t3tools/client-runtime/environment";
+import { RotateCcwIcon } from "lucide-react";
+import { useEffect, useMemo, useRef, useState } from "react";
+
+import { sortScopedProjectsForSidebar } from "../components/Sidebar.logic";
+import { Button } from "../components/ui/button";
+import { Empty, EmptyDescription, EmptyHeader, EmptyTitle } from "../components/ui/empty";
+import { SidebarInset } from "../components/ui/sidebar";
+import { useNewThreadHandler } from "../hooks/useHandleNewThread";
+import { useProjects, useThreadShells } from "../state/entities";
+
+export default function ExistingProjectIndex() {
+  const projects = useProjects();
+  const threads = useThreadShells();
+  const handleNewThread = useNewThreadHandler();
+  const startingRef = useRef(false);
+  const [startState, setStartState] = useState({ failed: false, retryRequest: 0 });
+  const mostRecentProject = useMemo(
+    () => sortScopedProjectsForSidebar(projects, threads, "updated_at")[0] ?? null,
+    [projects, threads],
+  );
+
+  useEffect(() => {
+    if (mostRecentProject === null || startingRef.current) return;
+    startingRef.current = true;
+    void handleNewThread(scopeProjectRef(mostRecentProject.environmentId, mostRecentProject.id), {
+      replace: true,
+    }).catch(() => {
+      startingRef.current = false;
+      setStartState((state) => ({ ...state, failed: true }));
+    });
+  }, [handleNewThread, mostRecentProject, startState.retryRequest]);
+
+  return startState.failed ? (
+    <SidebarInset className="h-dvh min-h-0 overflow-hidden overscroll-y-none bg-background text-foreground">
+      <Empty className="flex-1">
+        <EmptyHeader className="max-w-md">
+          <EmptyTitle className="text-foreground text-xl">Couldn’t start a new thread</EmptyTitle>
+          <EmptyDescription className="mt-2 text-sm text-muted-foreground/78">
+            The project is still available. Try opening the draft again.
+          </EmptyDescription>
+          <div className="mt-5 flex justify-center">
+            <Button
+              size="sm"
+              onClick={() => {
+                setStartState((state) => ({
+                  failed: false,
+                  retryRequest: state.retryRequest + 1,
+                }));
+              }}
+            >
+              <RotateCcwIcon className="size-4" />
+              Try again
+            </Button>
+          </div>
+        </EmptyHeader>
+      </Empty>
+    </SidebarInset>
+  ) : null;
+}
